% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agricultural_suitability.R
\docType{data}
\name{ramankuttyetal2002}
\alias{ramankuttyetal2002}
\alias{ramankutty}
\title{Ramankutty et al. (2002) Index for Global Agricultural Suitability}
\format{
##
A raster formatted as a stars object, unprojected EPSG:4326 (WGS84).
\describe{
  \item{cellvalue}{Agricultural suitability between 0 and 1.}
}
}
\source{
Ramankutty, Navin, Jonathan A. Foley, John Norman, and Kevin McSweeney. "The Global Distribution of Cultivable Lands: Current Patterns and Sensitivity to Possible Climate Change." Global Ecology and Biogeography 11, no. 5 (2002): 377-392. https://doi.org/10.1046/j.1466-822x.2002.00294.x.
}
\usage{
ramankutty
}
\description{
Agricultural suitability index scaled between 0 and 1.
The dataset describes the fraction of each 0.5-degree grid cell occupied by cultivated land at a global level.
}
\examples{
if (requireNamespace("stars", quietly = TRUE)) {
  library(stars)
  data("ramankuttyetal2002")

  plot(ramankutty)
 }
}
\keyword{datasets}
