% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agricultural_suitability.R
\name{ramankuttyetal2002}
\alias{ramankuttyetal2002}
\title{Ramankutty et al. (2002)}
\format{
##
A raster, unprojected
\describe{
  \item{country}{Country name}
  \item{iso2, iso3}{2 & 3 letter ISO country codes}
  \item{year}{Year}
  ...

}
}
\source{
Ramankutty, Navin, Jonathan A. Foley, John Norman, and Kevin McSweeney. "The Global Distribution of Cultivable Lands: Current Patterns and Sensitivity to Possible Climate Change." Global Ecology and Biogeography 11, no. 5 (2002): 377-392. https://doi.org/10.1046/j.1466-822x.2002.00294.x.
}
\description{
Agricultural suitability
}
